remove routes and added it to api
