+++ api.go
@@ -41,3 +41,11 @@ func (s *API) Create() {
 	log.Println("iam service running on port", s.listenAddr)
 	http.ListenAndServe(s.listenAddr, router)
 }
+
+func (s *API) IamRoutes(r *mux.Router) {
+	userPrefix := r.PathPrefix("/users").Subrouter()
+	r.HandleFunc("/auth", s.Auth).Methods(http.MethodPost)
+	r.HandleFunc("/logout", s.Logout).Methods(http.MethodDelete)
+
+	userPrefix.HandleFunc("/{id}", UserDetails).Methods(http.MethodGet)
+}
